requirements.txt
Flask==3.0.3
praw==7.7.1
python-dotenv-1.0.1